QUASAR: Add check for SIGSEGV and attempt to remove hhconfig.lock
Every time autoconfig is run, an attempt is made to remove the lockfile - just
for those strange cases (like SIGSEGV) where the lockfile isn't taken care of 
within the autoconfig executable. Added a check for SIGSEGV in Quasar as 
well, s.t. if it does happen one can be aware within the server console.