QUASAR: Changing startNovad message to use everyone.now.StartNovad
Of all the buttons on Pulsar, Start Classification was the one direct-message
button having problems. A message would be sent to start Novad, but for 
some reason the Quasar response to the message would start Novad, but 
fail to connect. Just calling the same method as the Start Classifier button
on Quasar, works now. May go through and switch over all of the message 
responses to do the same.