MULT: Creating config variable for iplist file path
It was hardcoded in two places, made it a config variable s.t. the hard
coded paths could be replaced with function calls. Removed redundant 
Logger initialization. Initializing dhcplist variable with Config::Inst() to 
make sure the Config constructor is run and the variable doesn't require a 
hard-coded path.

close #637