NOVAD: Send message to UIs to clear whitelisted suspect
This section was commented out with a TODO above, and was still
referencing the suspect table by IP instead of by SuspectIdentifiers. 
Modified it to use the SuspectIdentifiers vector returned by GetAllKeys
to erase whitelisted suspects from the table, and then notify the UIs so that
they might update their requisite presentation elements as well. Also, 
commented out a reference to everyone.now.SuspectCleared(s) from 
Quasar, as this function no longer exists and a type error was being thrown
every time a suspect was cleared.