AUTOCONFIG: Add LOG call to obviate TODO
When autoconfig is trying to set the default TCP behavior, that function can
return false, indicating that either the defaultBehavior passed to 
SetTCPDefaultBehavior was not within our enum or an assignment failed. 
Added a LOG call in the conditional watching that function s.t. the user 
can be notified if it were to fail.

close #652