LIB: Fix TODO regarding upde of haystack % contacted if haystack IPs change
Now changes in IP addresses of the honeypots will keep existing haystack
percent contacted data if possible. There are a lot of edge cases and 
questions regarding what to do to stats regarding haystack nodes contacted
if haystack IPs change, but this at least makes it so honeyd should be able
to update the IP address of a single node without novad resetting the
haystack percent contacted to 0 (but honeyd is currently lacking the 
functionality to change IPs of honeypots when running, so it's sort ofmoot).

Ticket #674