QUASAR: Checkboxes now activate slideUp/slideDown
Reworked the checkboxes for activation of a certain feature to act
and be seen a little more nicely. Now, a click of a checkbox shows or hides
a panel containing the relevant other fields for that feature.

Not going into integration until I test that the new rsyslog stuff (tcp/udp 
switch #733) is working as intended.

#733
#734