QUASAR: Duplicitous hack to fix rsyslog problem
Easiest way to do it, may not be pretty but the only other way to fix this 
is to turn almost everything in logger and the UIs on its head in a confusing
way...i.e. EMERGENCY is now 0, so 6+ (which is our current email alert level) 
becomes 1-. Not only is this a big change, but it's a pretty arbitrary change
to make, as it deviates from what, I think, would be expected by the user.
Anyways, this passes the Wireshark test for now, I'm going to go through
and try to test this on OpenSIEM to see if it all aggregates correctly.