LIB: Fix bug where doppelganger interface string blank and crashing honeyd
Actually two bugs. One was that the interface string was blank because
the Config.h file had some changes and the GetDoppelInterface method
was giving back the wrong variable, one that was never set.

The other problem was that editing the interface and ip of the doppelganger
node in the node editing page wasn't changing the values in NOVAConfig.txt. 

Ticket #749