CERES: Make server send all suspects
Works, but takes like 2 tries before it starts working because of
the way the suspectTable is divided between Nova and the web
C++ stuff. Need to find a way to have the suspects in the server
side before a call is made to get them. Also need to override 
notifyDataSetChanged to sort the list before updating the 
adapter.