CERES: Fix some macro usage, remove debug print, get suspects first time
Was using 1 and 0 for setVisibility, turns out 0 is Visible and 0x40000000 is
Invisible, so...yeah. Fixed that. Made the wait for the network timeout 
longer, so the suspects get loaded the first time. Might just make it wait
until the message comes in, no matter how long, but until I can test it on
an actual device that isn't the AVD I'll leave it as it is. Removed some
unused imports, implemented a custom comparator for sorting the Grid 
by classification.