CERES: Fixed configuration changed problem
Before, when you changed from portrait to landscape, onCreate
would get called again, which would parse the Xml from the 
StringReader in CeresClient. Well, turns out when you pass
the StringReader to the PullParser, it consumes whatever is
inside the StringReader, leading subsequent intializations of the 
parser to have no data. Fixed this, now on the grid page (and
presumably everywhere else) changing from portrait to 
landscape and vice-versa shouldn't throw any errors due to an 
empty StringReader