CERES: Many updates
Don't change activities when connection with server is lost.

There was a problem where going to the DetailsActivity would
always trigger the dialog saying that no data was fixed. should be
good now.

Overhauled threading behavior for GridActivity, some places the
thread wasn't getting interrupted, not good.

Making use of backstack now -- transition from Details to Grid 
works a little better, though I still need to find a way to have the
grid be maintained between the activity transitions; it seems like 
it should now that the flags for the Intents were changed around,
but it's re-triggering the thread every time. Still needs some work.

Login stuff is stored now; if it has login data it will just attempt to
connect right away. Need to monkey around with that dialog 
being cancellable, in case the user wants to connect to another
server or something.