MULT: Rewrite mail daemon in perl, other mods
The pyinotify library was too slow to catch all the notifications
that novad was sending, even in the reinstall-debug startup 
stuff. Rewrote it in perl, seems better now. Have yet to try it with
a more intense loggin regimen (like Nova getting hit with FAST) 
but if the syslog settings stay at the defaults they have on install, 
it might just be good enough. Now to test the sendmail
integration with cron...lol
Also had to add install lines for two perl modules that I need for 
the maildaemon to work right, Linux::Inotify2 and 
File::ReadBackwards. Added them to nova_init and 
the install helper script, so just need to add them to readme and 
all bases should be covered.