QUASAR package.json: set dependencies to newest versions of npm modules
Had a few side effects, like escaping changing in jade and express 3 having
some minor changes with the way you initialize the server. We had some 
modules locked in old versions because of some problems with Dojo AMD
conflicting with socket.io/nowjs,

https://github.com/LearnBoost/socket.io-client/issues/490

But now that dojo is gone, not an issue and we can go ahead and use the
newest versions of everything in the npm repos.