QUASAR: Add color differentiation between pie chart wedges
Rather than assigning colors independently randomly, assign colors according
to a random value within a window that is far enough away from the last color
so as to provide contrast.

This avoids the problem of getting two randomly generated colors that are
nearly identical, and having a difficult time telling them apart visually.