#820
This was just a side-effect of us using the /configureNovaSave stub for both
pages; obviously, when you save on the advanced page, there is no SMTP
INTERVAL dropdown, which leads that variable to be undefined in the 
req object, which leads to the error. Though the error doesn't stop anything
from saving (as we don't return after it happens), it should be fixed by
the attached changes.